+++ package/src/components/AudioPlayer/Context/TimeContext.ts
@@ -3,6 +3,7 @@ import { createContext } from "react";
 export interface TimeContext {
   currentTime: number;
   duration: number;
+  seekRequestKey: number;
 }
 
 export const timeContext = createContext<TimeContext | null>(null);