+++ package/src/components/AudioPlayer/Context/dispatchContext.ts
@@ -18,7 +18,8 @@ export type AudioContextAction<
   TInterfacePlacementLength extends number = typeof defaultInterfacePlacementMaxLength
 > =
   | { type: "NEXT_AUDIO" }
-  | { type: "PREV_AUDIO" }
+  | { type: "PREV_AUDIO"; currentTime: number }
+  | { type: "SEEK"; time: number }
   | { type: "UPDATE_PLAY_LIST"; playList: PlayList }
   | { type: "SET_AUDIO_STATE"; audioState: Partial<AudioState> }
   | {